backend/requirements.txt
fastapi
uvicorn
browser-use
langchain-google-genai
firebase-admin
python-dotenv
